+++ src/FAFS.Application/FAFSApplicationAutoMapperProfile.cs
@@ -1,13 +1,22 @@
 using AutoMapper;
+using FAFS.Destinations;                         // Entidad
+using FAFS.Application.Contracts.Destinations;  // DTOs
 
-namespace FAFS;
-
-public class FAFSApplicationAutoMapperProfile : Profile
+namespace FAFS
 {
-    public FAFSApplicationAutoMapperProfile()
+    public class DestinationApplicationAutoMapperProfile : Profile
     {
-        /* You can configure your AutoMapper mapping configuration here.
-         * Alternatively, you can split your mapping configurations
-         * into multiple profile classes for a better organization. */
+        public DestinationApplicationAutoMapperProfile()
+        {
+            // Mapear de entidad a DTO
+            CreateMap<Destination, DestinationDto>()
+                .ForMember(dest => dest.Latitude, opt => opt.MapFrom(src => src.Coordinates.Latitude))
+                .ForMember(dest => dest.Longitude, opt => opt.MapFrom(src => src.Coordinates.Longitude));
+
+            // Mapear de CreateDestinationDto a entidad 
+            CreateMap<CreateDestinationDto, Destination>()
+                .ForMember(dest => dest.Coordinates,
+                           opt => opt.MapFrom(src => new Coordinates(src.Latitude ?? "", src.Longitude ?? "")));
+        }
     }
 }
